move test files around
